rename: 사이드바 전용 아코디언 컴포넌트이므로 아예 Navigation 하로 옮김(#89)
